+++ nodejs/week2/documents.json
@@ -9,5 +9,10 @@
     "id": 2,
     "type": "doc",
     "value": "hello world"
+  },
+  {
+    "id": 3,
+    "type": "doc",
+    "value": "hello"
   }
 ]
